requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
pytz
psycopg2-binary
python-dotenv